backend/requirements.txt
flask
PyPDF2
python-docx
transformers
torch
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
